requirements.txt
langchain==0.0.222
chromadb==0.3.26
tiktoken==0.4.0
click==8.1.3
python-telegram-bot==20.4
runpod==1.3.4
python-dotenv==1.0.0
